+++ resources/views/admin/block/update.php
@@ -11,13 +11,16 @@ declare(strict_types=1);
 
 use Hirtz\Cms\Models\Block;
 use Hirtz\Cms\Modules\Admin\Widgets\Forms\BlockActiveForm;
+use Hirtz\Cms\Modules\Admin\Widgets\Navs\BlockActionDropdown;
 use Hirtz\Cms\Modules\Admin\Widgets\Navs\BlockHeader;
 use Hirtz\Cms\Modules\Admin\Widgets\Navs\BlockSubmenu;
 use Hirtz\Skeleton\Web\View;
 use Hirtz\Skeleton\Widgets\Forms\FormContainer;
 
 echo BlockHeader::make()
-    ->model($block);
+    ->model($block)
+    ->content(BlockActionDropdown::make()
+        ->model($block));
 
 echo BlockSubmenu::make()
     ->model($block);